gister/requirement.txt
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.4.0
numpy==1.20.1
Pillow==8.1.2
pydub==0.25.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytube==10.6.0
six==1.15.0
SpeechRecognition==3.8.1
typing-extensions==3.7.4.3
wordcloud==1.8.1
